+++ CHANGELOG.md
@@ -26,8 +26,12 @@
     - iMac7,1 - 18,x
     - MacPro3,1 - 6,1
       - MacPro3,1 can only boot with 4 cores max currently
+      - 8 cores can be re-enabled for older OSes in the GUI:
+        - Settings -> Build -> MacPro3,1/Xserve2,1 Workaround
     - Xserve2,1 - 3,1
       - Xserve2,1 can only boot with 4 cores max currently
+      - 8 cores can be re-enabled for older OSes in the GUI:
+        - Settings -> Build -> MacPro3,1/Xserve2,1 Workaround
   - Excludes the newly dropped MacBookAir8,x series.
     - No estimate can be given when support will be added.
   - For non-Metal graphics, Photos app will be broken.